bugs/feat: finally make dipole orthogonal in p-q
holy crap this has been a week of hunting the smallest bugs and chasing down the most miniscule of errors

Dipole looks great now, for realz this time. Works in 1 & 3 dimensions

This commit makes baselats_down a global var, and changes how the corners are put in. There are no longer any zeros, nans, etc.

note, some configurations of the grid *can* result in nan's at the poles, but these are rare

- code beforte this commit would have the last field line shifted one q-step lower than it should be. this makes everything even & makes sure the centers and corners are lined up.